project_requirements.txt
annotated-doc==0.0.4
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
colorama==0.4.6
decord==0.6.0
einops==0.8.2
ffmpeg-python==0.2.0
filelock==3.25.2
fsspec==2026.2.0
ftfy==6.1.1
future==1.0.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.12
iopath==0.1.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
opencv-python==4.13.0.92
packaging==26.1
pillow==12.1.1
portalocker==3.2.0
psutil==7.2.2
pycocotools==2.0.11
Pygments==2.20.0
pywin32==311
PyYAML==6.0.3
regex==2026.4.4
rich==15.0.0
safetensors==0.7.0
-e git+https://github.com/facebookresearch/sam3.git@967fdd651f71ca14949122fed4c918a778ca9334#egg=sam3&subdirectory=..\..\sam3
setuptools==70.2.0
shellingham==1.5.4
sympy==1.14.0
timm==1.0.26
torch==2.11.0+cu128
torchvision==0.26.0+cu128
tqdm==4.67.3
triton-windows==3.6.0.post26
typer==0.24.1
typing_extensions==4.15.0
wcwidth==0.6.0
